+++ common/src/main/java/com/github/dumann089/theatricalextralights/blockentities/MiniScanGobosBlockEntity.java
@@ -273,7 +273,6 @@ public class MiniScanGobosBlockEntity extends ExtraLightsLightBlockEntity
 
     @Override
     public int getColour() {
-        if (profile.isActive() && profile.isWheelRotating()) return profile.colour(0f);
         return ((getRed() & 0xFF) << 16) | ((getGreen() & 0xFF) << 8) | (getBlue() & 0xFF);
     }