+++ common/src/main/java/com/github/dumann089/theatricalextralights/blockentities/MovingVL2CBeamsBlockEntity.java
@@ -279,7 +279,6 @@ public class MovingVL2CBeamsBlockEntity extends ExtraLightsLightBlockEntity impl
 
     @Override
     public int getColour() {
-        if (profile.isActive() && profile.isWheelRotating()) return profile.colour(0f);
         return ((getRed() & 0xFF) << 16) | ((getGreen() & 0xFF) << 8) | (getBlue() & 0xFF);
     }