+++ common/src/main/java/com/github/dumann089/theatricalextralights/blockentities/SpotXtremeGoboBlockEntity.java
@@ -275,7 +275,6 @@ public class SpotXtremeGoboBlockEntity extends ExtraLightsLightBlockEntity
 
     @Override
     public int getColour() {
-        if (profile.isActive() && profile.isWheelRotating()) return profile.colour(0f);
         return ((getRed() & 0xFF) << 16) | ((getGreen() & 0xFF) << 8) | (getBlue() & 0xFF);
     }