+++ common/src/main/java/com/github/dumann089/theatricalextralights/blockentities/VL6CGoboBlockEntity.java
@@ -274,7 +274,6 @@ public class VL6CGoboBlockEntity extends ExtraLightsLightBlockEntity
 
     @Override
     public int getColour() {
-        if (profile.isActive() && profile.isWheelRotating()) return profile.colour(0f);
         return ((getRed() & 0xFF) << 16) | ((getGreen() & 0xFF) << 8) | (getBlue() & 0xFF);
     }